Adopt OpenJDK: Change the download script from oracle jdk distribution to openJDK Distribution (because otherwise an oracle.com login would be required and that script was a common point of failure in history anyway, because older jdk's got removed, urls changed, ...).
This script has to be updated for jdk11, because the naming scheme changed a bit.
Currently it is untested (it generates everything, but it could be that something is missing and thus the jdk's executables are in the wrong folder, so a little "mv" instruction has to be added), so please try it out!